+++ parselcli/config.py
@@ -18,6 +18,7 @@ CACHE_DIR = XDG_CACHE_HOME / APP_NAME
 DEFAULT_CONFIG = {
     # default processors that are activated on startup
     "color": True,
+    "vi_mode": False,
     "warn_limit": 5000,
     "initial_input": [],
     "history_file_css": str(CACHE_DIR / "history_css"),